fix bai 7

+++ src/Bai7.java
@@ -1,6 +1,6 @@
 public class Bai7 {
     public static void tinhGiaiThua(int n){
-        int s=1;
+        double s=1;
         for (int i=1;i<=n;i++){
             s*=i;
         }
